+++ docs/changes.rst
@@ -14,6 +14,7 @@ Released on ? 2017.
 * Improved rendering of inconsistent translations.
 * Added support for opening source files in local editor.
 * Added support for condiguring visual keyboard with special characters.
+* Improved screenshot management with OCR support for matching source strings.
 
 weblate 2.11
 ------------